HH84: update types
