+++ src/events/events/email/email-registration-links-to-students.event.ts
@@ -1,5 +1,5 @@
-import { Student } from '../../../students/entities';
+import { StudentEntity } from '../../../types';
 
 export class EmailRegistrationLinksToStudentsEvent {
-  students: Student[];
+  students: StudentEntity[];
 }
