--- src/types/email/index.ts
@@ -1,1 +1,0 @@
-export * from './email';
