+++ src/types/index.ts
@@ -8,4 +8,3 @@ export * from './hr';
 export * from './student-grades';
 export * from './error';
 export * from './student-profile';
-export * from './email';
