+++ src/utils/mail/mail-details.ts
@@ -7,3 +7,13 @@ export enum MailTemplate {
   REGISTER = 'register',
   RESTORE_PWD = 'restore-pwd',
 }
+
+export interface SendEmailOptions {
+  email: string;
+  template: MailTemplate;
+  subject: MailSubject;
+  context: {
+    [name: string]: unknown;
+  };
+  from?: string;
+}
